#16 금요일 FollowUp - 2 회색 화면 출력이긴 함

+++ Inho/InhoDx2D/Project/Engine/CAnim.cpp
@@ -25,7 +25,7 @@ void CAnim::finaltick()
 	if (m_vecFrm[m_CurFrmIdx].Duration < m_AccTime) {
 		++m_CurFrmIdx;
 		if (m_vecFrm.size() <= m_CurFrmIdx) {
-			m_CurFrmIdx = m_vecFrm.size() - 1;
+			m_CurFrmIdx = (int)m_vecFrm.size() - 1;
 			m_bFinish = true;
 		}
 		m_AccTime = 0.f;
